Don't nest imenu to arbitrary depths
Even though helm-imenu seems to support arbitrary levels of nesting,
imenu itself (i.e. imenu-add-to-menubar) does not.

Revert to the old hierarchy in which the first level is used for the
leaf symbol type, however, keep the hierarchical data in the form of
using the "fully-qualified" (a.b.c) name for the imenu item titles.